third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-transition-transform-function-expected.txt

This is a testharness.js-based test.
FAIL A custom property of type <transform-function> can yield a CSS Transition assert_equals: Element has the expected animated value expected "translateX(150px)" but got "translateX(200px)"
Harness: the test ran to completion.
